requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
torch==2.2.2 
torchvision==0.17.2 
torchaudio==2.2.2
